# backend/app/services/github/schema.py
from enum import Enum
from typing import List, Dict, Any, Literal, Optional

from pydantic import BaseModel, Field


class ChatRequest(BaseModel):
    message: str
    context: Optional[Dict[str, Any]] = None


class ChatResponse(BaseModel):
    response: str
    code_snippets: Optional[List[Dict[str, Any]]] = None
    source_files: Optional[List[Dict[str, str]]] = None


class RepositoryStatusResponse(BaseModel):
    status: str
    file_count: Optional[int] = None
    indexed_at: Optional[str] = None
    message: Optional[str] = None


class AgentRequest(BaseModel):
    """Generic request for any agent that takes a simple prompt."""

    prompt: str
    context: Optional[Dict[str, Any]] = None


class AgentResponse(BaseModel):
    """Generic response for an agent that returns a simple text response."""

    response: str
    agent_id: str
    execution_time: float
    metadata: Optional[Dict[str, Any]] = None


class TaskType(str, Enum):
    """The types of tasks our agent can create. Start with one, add more later."""

    PHONE = "phone_task"
    RESEARCH = "research_task"
    IMAGE_GENERATION = "image_generation_task"


class NewCardData(BaseModel):
    """
    Defines the structure for a single task card, now with dependency tracking.
    """

    card_id: str = Field(
        ...,
        description="A unique temporary ID for this card within the response (e.g., 'task-1').",
    )
    title: str
    description: str
    task_type: TaskType
    status: Literal["todo", "doing", "done"] = "todo"
    dependencies: list[str] = Field(
        default_factory=list,
        description="List of card_ids this card depends on.",
    )


class NewCardAgentResponse(BaseModel):
    """The final response, containing a list of generated cards."""

    card_data: list[NewCardData]
    agent_id: str
    execution_time: float
    metadata: Dict[str, Any]


class ImageGenerationRequest(BaseModel):
    """Request model for generating an image based on a text prompt."""

    prompt: str = Field(
        ...,
        description="A descriptive prompt for the image to be generated.",
        min_length=10,
    )


class ImageGenerationResponse(BaseModel):
    """
    Response model containing the generated image as a Base64 string.
    """

    image_base64: str = Field(
        ..., description="The generated image, encoded as a Base64 string."
    )
    model_id: str = Field(..., description="The model used for generation.")